Refactor: Add express routers to separate app.js into different files
This change increases readability of app.js and makes use of express's tools to improve the structure of the project.
